Refactored to loosen plan recognition from concrete BATs.
There are two new typeclasses, obs_bat and pr_bat. The former is a
subclass of bat, the latter of the former.
Thus a BAT should basically implement the pr_bat typeclass in order
to be plan-recognizable.

The plan recognition and obs modules just rely on these typeclasses
instead of depending on the concrete implementations.
This should allow us to easily switch between different BATs by just
chaning the code in, e.g., cars_server.m.

Disclaimer:
1. It's untested, it just compiles.
2. Some cleanup is really necessary (like code reformatting, removing
   superfluous imports).
3. Probably some more refactoring needs to be done.