Refactored observations as part of BATs.
This was a major refactoring. Some slight renaming is left.

I had a great problem with (a pretty bad) compiler error. My fault was
to make `Stream' a parameter of the typeclass obs_source, but in the
planrecog module, the respective parameter was filled with an int
somewhere.
Actually, this was correct; so now there is no such parameter called
`Stream'. Instead, stream == int is used everywhere.