Procedure calls now nullary higher order functions.
Before, procedure calls were types and thus explicit parameters of the BAT.
The body of a procedure call was obtained with the proc/2 predicate
defined in the BAT.

We could probably do similar for stochastic actions by replacing each
stochastic action with a unary higher order function which, given the
current situation, returns the `random' outcome.
This would replace the current random_outcome/3 predicate.
(Note that random_outcome/3 actually is not random; the random seed is
somewhere encoded in the situation term and then random_outcome/3 just
deterministically computes a primitive outcome action.)

Maybe we could even implement nondeterministic pick this way?

Finally, one might want a layer on-top of this procedure-less and
stochastic-action-less BATs that allows for cleaner programs.
This layer could be a super-typeclass of BAT that again requires a type
for procedure calls and stochastic actions and then replaces their
occurrences with respective higher-order terms.