Removed coin-Osi-gc from Makefile.
We don't use the standalone solver (that is, COIN-OR embedded linked to
the libosi.so compiled from osi.m). Instead, we use the remote solver
that communicates with the lp-solver binary via AF_UNIX sockets.

To the standalone solver we needed to change the Mercury compiler's
Boehm GC. Probably this led to the "finalization cycle" warnings that
occurred, for example, when we used cars-server multiple times.