Testing RSTC BAT.
Problem with tailgate program is that after accel(v_c/v_b),
TTC(c,b) is undefined. After a subsequent wait(_), also
NTG(c,b) is undefined because it needs TTC(c,b) to determine NTG(c,b)
in general.

This should be fixable, because in this scenario, NTG(c,b) should
simply remain the same after subsequent wait(_) actions.

One way would be to add infinity to arithmetic. If that's done properly,
the successor state axioms could remain as they are, don't they?

Maybe a simpler solution is to say that if NTG(c,b) is defined but
TTC(c,b) is not, then the latter can be assumed to be infinity (and the
other way around). These cases could be considered in the successor
state axioms.