prGolog: Implemented pick. Added tree container.
A pick operator consists of an (existential) data type T and an unary
function that returns a program for each T. Furthermore there is an
initial value of type T and a binary successor function F that returns
for a given value of type T and a quality measure value the next value
of type T to be picked.

New values are picked until F reaches a fixed point.
The idea behind that is that F can thus search the best pick value.

It's not easy to decide where to do this optimization. In the formal
semantics, the `next' predicate simply `returns' all possible
decompositions. That obviously not implementable.

Now we use higher-order functions to compute them lazily.
This higher-order function is unary, the type of the parameter is a
new type called pickbest. This is again a higher-order function.
The (currently only) implementation is the one that computes the
sequence
  X0, F(X0), F(F(X0)), F(F(F(X0))), ...
until a fixpoint, say X, is reached. Then it returns P(X).

We had to do some tricks because Mercury doens't allow universally
quantified type variables in higher-order function types. The main
part of this trick is the pickspec type, which has existentially
type T.