Fixed pick operator.
The pick operator
   pick(..., prog1) ; prog2
(or other operatores for ;) maximized only over prog instead over prog1 but
ignored prog2.

I fixed this now. The best value must be picked wrt the set of
decompositions.
That is, while the lazy nodes were nodes of a tree of before, they're
now nodes of a tree of pseudo_decomp's and decomp's.
Furthermore, the argument of the lazyness functions isn't a pickbest
function anymore but a next_func, a starting value, and a value_func.
The actual back coupling and optimization is now done in tree.force/2.

Implementing this was really hard by the way.

Unfortunately for some reason the compiler doesn't like tree insts of as
much as before. Why is that? Check on this.