Fixed: used wrong constant for minimal float value in prgolog!
float.min is the the minimal *positive* number.
-1.0 * float.max is the smallest float.

That led to different behavior between the new value/3 implementation
that uses tree.map_reduce/3 where list/tree.fold/3 was used before.

Also fixed a bug in the unit test which expected {0.0, 0} as value for a
program that consists of three actions for a lookahead of 10. The right
expected value is {0.0, 10}.
The program behaved incorrectly (returning the old, wrong expected
value) with the float.min-bug and when list/tree.foldl/3 is used in
value/3.
With the float.min-bug fixed, the program behaves correctly even with
when tree.foldl/3 is used (but currently I prefer the version with
tree.map_reduce/3 anyway).

Thanks for 1.5 hours less sleep.