Added random_generator typeclass; PSO thus more generic; multi-var pick.
* New module: util/util.rand.m wraps some random number generator
  functionality. Provides this random_generator typeclass with
  implementations for int, float, tuples and triples.
* New module: util/util.vector_space.m, util/util.vector_space.impl.m
  define and implement a simple and very small vector space typeclass.
* PSO now uses the two aforementioned typeclasses. Thus it no longer
  optimizes functions float -> U, but T -> U such that T implements the
  random_generator and a vector_space typeclasses.

This allows to pick and optimize for tuples. For example we can pick
two accelerations at once.
I hoped this approach would be faster than two nested PSOs, which are
pretty slow because the function (basically = prgolog.value/2) is so
expensive.
However, this didn't work out so well (until now): we need a very
large lookahead so that the second picked variable is reasonable
optimized.

We keep the aforementioned changes/additions anyway because they might
be useful later.