Changed reward/1 of end(_, _) actions.
The theory before was to return 2 * sitlen(S) in order to pull back
end(_, _) actions. Unfortunately this doesn't work in practice because
the interpreter sees the end(_, _) at the end of the lookahead with all
its sweet reward. Therefore it simply executes the candidate program
just to get the tempting end(_, _) action.

For that reason we returned the 0.0 for some time now and it worked
fine.

However, this leads to some observations not being explained although
they could be explained! (They weren't executed after the end(_, _)
action because the planrecog.m module stops execution when the candidate
program is final (it replaces all observations with nil and checks
final/2)).

Now the reward for end(_, _) is 2.0 * sitlen(S) if the previous action
was an observation, and otherwise 0.0.
This makes end(_, _) untempting for the interpreter for a long time.
Only when the last non-end(_, _) action is covered by an observation, we
make the end(_, _) action worth striving for. This avoids that the
interpreter prefers the wait(_) of an observation and then fails with
the match(_, _). However, since the reward for the end(_, _) action is
2.0 * sitlen(S), the interpreter will prefer observations if it can
execute them.

By the way this commit also fixes that for some reason we had removed
the comment symbol in front of some wait_until action in the overtake
program which shouldn't have been removed.