Implemented heuristic search for decompositions.
The tree.m module now provides facilities to extract the maximum-valued
entries from the tree. Besides the value function one needs a heuristic
which does not underestimate the value (otherwise it's not optimal).

Trees now have lazy nodes, whose content is computed upon evaluation,
and sprouts, whose optimal content can be computed by force/2 or lazily
using load/2.

In the first step, the whole tree is valued with the heuristic. Sprouts
are valued with their X0 value. Note that the heuristic for X0 must be
greater than or equal to the optimally value to keep admissibility.
then load/2 is called to lazily grow the sprouts optimally.
The maximally estimated nodes are extracted and their value is computed.
This is repeated until the next-best estimated node has a lower
estimated value than the value of best-so-far node.